Move some logic
